Creacion Servicios Libro,Autor, y Editorial 28-12-22

+++ Ejercicios/Libreria/src/Logica/DAO/AutorDAO.java
@@ -28,7 +28,7 @@ public class AutorDAO extends DAO<Autor> {
         }
     }
 
-    public void eliminarAutorPorId(Long id) {
+    public void eliminarAutorPorId(Integer id) {
         try {
             conectar();
             Autor a = em.find(Autor.class, id);
